Code-freezed/MPAI-CAE-ARP Audio Analyser/pyproject.toml
[tool.poetry]
name = "audio-analyzer"
version = "1.0.0"
description = "MPAI CAE-ARP Audio Analyser"
repository = "https://gitlab.dei.unipd.it/mpai/audio-analyzer.git"
authors = ["Matteo Spanio <dev2@audioinnova.com>"]
license = "GPLv3"
readme = "README.md"
packages = [
    { include = "audio_analyzer", from = "src" },
    { include = "ml", from = "src" }
]

[tool.poetry.scripts]
audio-analyser = 'audio_analyzer.cli:main'

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.4"    # imposed by mpai-cae-arp
rich = "^13.3.3"
pandas = "^2.1.0"
scikit-learn = "^1.2.2"
grpcio-tools = "^1.53.0"
mpai-cae-arp = "^0.5.0"
ffmpeg-python = "^0.2.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^6.1.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.261"
pytest = "^7.3.0"
pytest-xdist = "^3.2.1"
pytest-json-report = "^1.5.0"
jsonschema = "^4.19.0"
filetype = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
src = ["src", "tests"]
select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "UP",   # pyupgrade
]

[tool.pytest.ini_options]   # eventually overridden by pytest.ini files
minversion = "6.0"
addopts = "-ra -q -W error::RuntimeWarning --import-mode=importlib"
testpaths = ["tests", "conformance_tests"]
